requirements.txt
langchain==0.1.0
openai==1.3.0
faiss-cpu==1.7.4
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
unstructured==0.10.30
pdf2image==1.16.3
pytesseract==0.3.10 